backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
python-multipart==0.0.9
pymupdf==1.24.6
Pillow==10.4.0
modal-client==0.61.0
python-dotenv==1.0.1
